+++ kafka-consumer/src/main/java/by/volkov/consumer/vm/RateVM.java
@@ -5,6 +5,7 @@ import lombok.AccessLevel;
 import lombok.Data;
 import lombok.experimental.FieldDefaults;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
@@ -12,10 +13,10 @@ import java.time.LocalDateTime;
 public class RateVM {
 
     Long id;
-    Double sellRate;
+    BigDecimal sellRate;
     Currency sellIso;
     Integer sellCode;
-    Double buyRate;
+    BigDecimal buyRate;
     Currency buyIso;
     Integer buyCode;
     Integer quantity;
